+++ src/components/elements/BlockLink.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { withLink } from 'helpers'
+import { withLink } from 'helpers/hoc'
 import Box from './Box'
 
 const BlockLink = styled(Box)({
