+++ src/helpers/hoc/index.js
@@ -1,0 +1,4 @@
+import withLink from './with-link'
+import withSlug from './with-slug'
+
+export { withLink, withSlug }
