+++ src/helpers/hoc/with-slug.js
