+++ src/helpers/index.js
@@ -6,19 +6,13 @@ import get from './get'
 import createApiUrl from './create-api-url'
 import smoothScroll from './smooth-scroll'
 import serializeComponent from './serialize-component'
-import withSlug from './with-slug'
-import withLink from './with-link'
-import createCssState from './create-css-state'
 
 export {
   serializeComponent,
-  createCssState,
-  withSlug,
   smoothScroll,
   marshall,
   unmarshall,
   get,
-  withLink,
   formatDate,
   formatNumber,
   createApiUrl
